fix leader election bug introduced by #37.
I just found a bug in #37. When the leader is lost(mainly due to apiserver timeout), the program will stop watch events but will not exit，it will always hung there.
I modified this part of the logic in this pr. After the leader is lost, the program will exit the same as it received exit signal. Then a new pod will be scheduled to participate in the election.
Sorry for introducing this bug in # 37, hope this pr will fix it.